stimulator_ui/requirements.txt
Flask
PyQt5
socketio
pyserial